exercises route done

+++ controllers/workoutController.js
@@ -5,19 +5,20 @@ const Workout = require('../models/Workouts')
 // CREATE METHOD
 const createWorkout  =  async (req, res)=>{
     try {
-        const {name, day,  description} = req.body
+        const {name, day, description} = req.body
         if (!name || !day  ||  !description){
             return res.status(400).json({status:'error' , msg:'Plese  make sure all fields are are filled in. '})
         } 
 
         //AFTER PASSING THE VALLIDATION 
-const workout = await Workout.create({
-    name,
-    day,
-    description,
-    user: "68b007113d34e73b635da35b"
-})
-        res.status(200).json({msg:'All data was recevied successfully'})
+        const workout = await Workout.create({
+            name,
+            day,
+            description,
+            user: res.locals.user.id
+        })
+        // res.status(200).json({msg:'All data was recevied successfully'})
+        res.send(workout)
     } catch (error) {
         console.log(error)
         res.status(500).json({status:'error' ,msg:'An unexpected error occurred '})
@@ -31,9 +32,10 @@ const workout = await Workout.create({
 // get workouts 
 const getWorkouts = async (req, res)=> {
     try {
-        const workouts = await Workout.find({user:req.userId}).sort({createdAt:-1})
+        const workouts = await Workout.find({user: res.locals.user.id}).sort({createdAt:-1})
         res.status(200).json(workouts)
     } catch (error) {
+        console.log(err)
         res.status(500).json({status:'error' ,msg:'An unexpected error occurred '})
     }
 }
@@ -56,12 +58,14 @@ const getWorkout = async (req, res) => {
 
 // UPDATE METHOD
 const updateWorkout = async(req, res)=> {
-    
     try {
         // NOT FOUND TO UPDATE
         const workout  = await Workout.findById(req.params.id)
         if(!workout){
             return res.status(404).json({msg: 'Workout not found'})
+        }
+        if(!workout.user.equals(res.locals.user.id)) {
+            return res.status(401).json({msg: "Unauthorized Access!"})
         }
         const updated = await Workout.findByIdAndUpdate(
             req.params.id, //search for the workout that want to be edited
@@ -86,9 +90,11 @@ const deleteWorkout = async (req, res)=>{
         if(!workout){
             return res.status(404).json({msg:'Workout not found'})
         }
+        if(!workout.user.equals(res.locals.user.id)) {
+            return res.status(401).json({msg: "Unauthorized Access!"})
+        }
         const deleted = await Workout.findByIdAndDelete(req.params.id)
         res.status(200).json({msg:'Workout deleted successfully'})
-        
     } catch (error) {
         res.status(500).json({status:'error',msg:'not found'})
     }
